avrflash: rewrite to use own protocol, plus avrdude patch
STK500 may be fine, but it's just easier to write a simple driver that just
forwards AVR commands, instead of repacking data blocks into larger blocks
that have to be unpacked again.

So this is basically just accepting commands for the AVR and a
"reset high-low" command. Includes avrdude patch that implements this
programmer protocol, is basically STK500 just everything but the "universal"
command removed.

This also means that the programmer now supports everything that avrdude can
do.